fix multi round fixture
