+++ compiler/src/test/fixtures/ksp-multi-round/library/api/build.gradle.kts
@@ -7,20 +7,13 @@ plugins {
 anvil {
     useKsp(
         contributesAndFactoryGeneration = true,
-        componentMerging = true,
     )
 }
 
 dependencies {
     implementation(projects.processorApi)
     implementation(projects.di)
-    implementation(projects.library.api)
-    implementation(projects.library.impl)
-    implementation(libs.anvil.utils.annotations)
     implementation(libs.dagger)
 
     ksp(projects.processor)
-    ksp(libs.anvil.utils.compiler)
-
-    ksp(libs.dagger.compiler)
-} 
+}
