+++ compiler/src/test/fixtures/ksp-multi-round/library/api/src/main/kotlin/me/gulya/anvil/sample/SampleComponent.kt
@@ -1,0 +1,8 @@
+package me.gulya.anvil.sample
+
+@GenerateClass // This will trigger our KSP processor to generate SampleComponentGenerated
+interface SampleComponent {
+    fun interface Factory {
+        operator fun invoke(generated: SampleComponentGenerated): SampleComponent
+    }
+}
