+++ compiler/src/test/fixtures/ksp-multi-round/library/impl/src/main/kotlin/me/gulya/anvil/sample/DefaultSampleComponent.kt
@@ -4,16 +4,7 @@ import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
 import me.gulya.anvil.assisted.ContributesAssistedFactory
 
-@GenerateClass // This will trigger our KSP processor to generate SampleComponentGenerated
-interface SampleComponent {
-    fun interface Factory {
-        operator fun invoke(generated: SampleComponentGenerated): SampleComponent
-    }
-}
-
 @ContributesAssistedFactory(SampleScope::class, SampleComponent.Factory::class)
 class DefaultSampleComponent @AssistedInject constructor(
     @Assisted val generated: SampleComponentGenerated // This depends on the generated class
 ) : SampleComponent
-
-class SampleScope
